Restructuring for ease of use underway
